pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ana_p3"
version = "0.1.0"
description = "Package for project 3 in Advanced Numerical Algorithms Course"
readme = "README.md"
requires-python = "==3.13.*"
license = { text = "MIT" }
authors = [
    { name = "Philip Korsager Nickel", email = "s214960@dtu.dk" }
]

# All dependencies (runtime, docs, and dev)
dependencies = [
  # Runtime dependencies
  "numpy>=1.24.0",
  "matplotlib>=3.7.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "scipy>=1.11.0",
  "seaborn>=0.12.0",
  # Documentation dependencies
  "sphinx>=7.0.0",
  "pydata-sphinx-theme>=0.14.0",
  "numpydoc>=1.6.0",
  "sphinx-copybutton>=0.5.2",
  "sphinx-gallery>=0.14.0",
  "pillow>=10.0.0", # Required for sphinx-gallery image handling
  # Development dependencies
  "ruff>=0.1.0",
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "numba>=0.62.1",
  "gmsh>=4.15.0",
  "meshio>=5.3.5",
  "pyyaml>=6.0.3",
  "pyvista>=0.46.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]



